restructure typography components
